ticket related changes,add filter in city

+++ src/main/java/com/nice/service/impl/ProductServiceImpl.java
@@ -83,7 +83,7 @@ import com.nice.util.ExportCSV;
 
 /**
  * @author : Kody Technolab PVT. LTD.
- * @date : 29-Jun-2020
+ * @date   : 29-Jun-2020
  */
 @Service(value = "productService")
 @Transactional(rollbackFor = Throwable.class)
@@ -161,8 +161,11 @@ public class ProductServiceImpl implements ProductService {
 	@Autowired
 	private BusinessCategoryService businessCategoryService;
 
-	@Value("${default.product}")
-	private String defaultProduct;
+	@Value("${product.list.image}")
+	private String productListImage;
+
+	@Value("${product.detail.image}")
+	private String productDetailImage;
 
 	private static final String NON_VEG = "Non_Veg";
 
@@ -232,9 +235,15 @@ public class ProductServiceImpl implements ProductService {
 		product.setBusinessCategoryId(vendorDto.getBusinessCategoryId());
 		if (image != null) {
 			uploadImage(image, product);
+		} else {
+			product.setImage(productListImage);
+			product.setImageOriginalName(productListImage);
 		}
 		if (detailImage != null) {
 			uploadDetailImage(detailImage, product);
+		} else {
+			product.setDetailImage(productDetailImage);
+			product.setDetailImageOriginalName(productDetailImage);
 		}
 		productRepository.save(product);
 		LOGGER.info("addProduct executed successfully");
@@ -264,7 +273,7 @@ public class ProductServiceImpl implements ProductService {
 	private void updateImages(final MultipartFile image, final Product existingProduct, final Product product)
 			throws FileOperationException, ValidationException {
 		LOGGER.info("Inside image for Product : {}", product.getId());
-		if (image != null) {
+		if (image != null && !existingProduct.getImage().equals(productListImage)) {
 			deleteOldImage(existingProduct.getImage());
 			uploadImage(image, product);
 		} else {
@@ -277,7 +286,7 @@ public class ProductServiceImpl implements ProductService {
 	private void updateDetailImage(final MultipartFile detailImage, final Product existingProduct, final Product product)
 			throws FileOperationException, ValidationException {
 		LOGGER.info("Inside image for Product : {}", product.getId());
-		if (detailImage != null) {
+		if (detailImage != null && !existingProduct.getDetailImage().equals(productDetailImage)) {
 			deleteOldImage(existingProduct.getDetailImage());
 			uploadDetailImage(detailImage, product);
 		} else {
@@ -377,7 +386,7 @@ public class ProductServiceImpl implements ProductService {
 	/**
 	 * validation for add or update product
 	 *
-	 * @param productRequestDTO
+	 * @param  productRequestDTO
 	 * @throws NotFoundException
 	 * @throws ValidationException
 	 */
@@ -486,13 +495,12 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * listForAdmin==null means get product detail for admin listForAdmin==true means get product list for admin convert
-	 * entity to response dto
+	 * listForAdmin==null means get product detail for admin listForAdmin==true means get product list for admin convert entity to response dto
 	 *
-	 * @param product
-	 * @param listForAdmin
-	 * @param productParamRequestDTO
-	 * @param pincodeId
+	 * @param  product
+	 * @param  listForAdmin
+	 * @param  productParamRequestDTO
+	 * @param  pincodeId
 	 * @return
 	 * @throws NotFoundException
 	 * @throws ValidationException
@@ -591,8 +599,8 @@ public class ProductServiceImpl implements ProductService {
 		}
 
 		/**
-		 * Here flag is set to determine whether to make product visible to customer or not, if the setProductAvailable flag is
-		 * true then product should be made visible else that product should be shown as out of stock
+		 * Here flag is set to determine whether to make product visible to customer or not, if the setProductAvailable flag is true then product should be made
+		 * visible else that product should be shown as out of stock
 		 */
 		if (businessCategoryDto.getName().equalsIgnoreCase(Constant.BUSINESS_CATEGORY_GROCERY)) {
 			productResponseDTO.setProductAvailable(availableQty > 0);
@@ -635,9 +643,9 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * @param active
-	 * @param existingProduct
-	 * @param productId
+	 * @param  active
+	 * @param  existingProduct
+	 * @param  productId
 	 * @throws NotFoundException
 	 * @throws ValidationException
 	 */
@@ -667,7 +675,7 @@ public class ProductServiceImpl implements ProductService {
 	/**
 	 * check masters for activate product
 	 *
-	 * @param existingProduct
+	 * @param  existingProduct
 	 * @throws ValidationException
 	 * @throws NotFoundException
 	 */
@@ -697,8 +705,8 @@ public class ProductServiceImpl implements ProductService {
 
 		UserLogin userLogin = getUserLoginFromToken();
 		/**
-		 * In case of customer the userLogin might be anonymous user, resulting in no user login and hence if the userLogin is
-		 * null or the userLogin->entityType is customer then we will give records specific to customer
+		 * In case of customer the userLogin might be anonymous user, resulting in no user login and hence if the userLogin is null or the userLogin->entityType
+		 * is customer then we will give records specific to customer
 		 */
 		if (userLogin != null && UserType.VENDOR.name().equals(userLogin.getEntityType())) {
 			productParamRequestDTO.setVendorId(userLogin.getEntityId());
@@ -711,8 +719,8 @@ public class ProductServiceImpl implements ProductService {
 	/**
 	 * upload image of product
 	 *
-	 * @param image
-	 * @param product
+	 * @param  image
+	 * @param  product
 	 * @throws ValidationException
 	 * @throws FileOperationException
 	 */
@@ -832,8 +840,8 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * @param productImportDTOs
-	 * @param userId
+	 * @param  productImportDTOs
+	 * @param  userId
 	 * @return
 	 */
 	private List<ProductImportDTO> insertListOfProducts(final List<ProductImportDTO> productImportDTOs) {
@@ -905,10 +913,10 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * @param productImportDTO
-	 * @param brandId
-	 * @param vendor
-	 * @param productRequestDTO
+	 * @param  productImportDTO
+	 * @param  brandId
+	 * @param  vendor
+	 * @param  productRequestDTO
 	 * @throws ValidationException
 	 */
 	private void validateAndSetProductImport(final ProductImportDTO productImportDTO, final Long brandId, final Vendor vendor,
